refactor(visualization): break codeCharta.model import cycles
codeCharta.model (the leaf domain types) was reaching up into two higher
layers, forming type-only cycles:
- NameDataPair.content needed ExportCCFile -> codeCharta.model <-> codeCharta.api.model
- CcState.files needed FileState -> codeCharta.model <-> model/files/files

Fix by making the leaf stop importing upward:
- Move NameDataPair into codeCharta.api.model (next to ExportCCFile); repoint
  its 8 type importers.
- Move FileState/FileSelectionState into codeCharta.model; model/files/files
  re-exports them (natural files->model direction), so its 47 importers are
  untouched.

Both edges removed: dependency-cruiser warnings 96 -> 94, 0 errors, tests green.
