+++ visualization/app/codeCharta/features/loadFile/services/loadFile.service.ts
@@ -3,7 +3,7 @@ import { tap } from "rxjs"
 import { clone } from "../../../util/clone"
 import { CCFileValidationResult } from "../util/fileValidator"
 import { FileState } from "../../../model/files/files"
-import { NameDataPair } from "../../../codeCharta.model"
+import { NameDataPair } from "../../../codeCharta.api.model"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
 import { loadFilesValidationToErrorDialog } from "./loadFilesValidationToErrorDialog"
 import { enrichFileStatesAndRecentFilesWithValidationResults } from "../util/fileParser"
