+++ visualization/app/codeCharta/features/loadFile/services/loadInitialFile.service.ts
@@ -5,7 +5,8 @@ import { readCcState } from "../../../util/indexedDB/indexedDBWriter"
 import sample1 from "../../../assets/sample1.cc.json"
 import sample2 from "../../../assets/sample2.cc.json"
 import { ExportCCFile } from "../../../codeCharta.api.model"
-import { CcState, NameDataPair } from "../../../codeCharta.model"
+import { CcState } from "../../../codeCharta.model"
+import { NameDataPair } from "../../../codeCharta.api.model"
 import { FileState } from "../../../model/files/files"
 import { MetricQueryParemter } from "../../../state/effects/updateQueryParameters/metricQueryParameter"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
