+++ visualization/app/codeCharta/features/loadFile/util/fileValidator.spec.ts
@@ -5,7 +5,8 @@ import {
     TEST_FILE_CONTENT_INVALID_MINOR_API,
     TEST_FILE_CONTENT_NO_API
 } from "../../../mocks/dataMocks"
-import { CodeMapNode, NameDataPair, NodeType } from "../../../codeCharta.model"
+import { CodeMapNode, NodeType } from "../../../codeCharta.model"
+import { NameDataPair } from "../../../codeCharta.api.model"
 import packageJson from "../../../../../package.json"
 import { checkErrors, checkWarnings, ERROR_MESSAGES } from "./fileValidator"
 import { fileWithFixedFolders, fileWithFixedOverlappingSubFolders } from "../../../resources/fixed-folders/fixed-folders-example"
